pyproject.toml
[project]
name = "heard"
version = "0.1.0"
description = "Video speech-to-text transcription CLI tool"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15.0",
    "ffmpeg-python>=0.2.0",
    "faster-whisper>=1.1.0",
    "rich>=13.0.0",
]

[project.scripts]
heard = "heard.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/heard"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
